+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - accepts all forms of `cd` type inputs, aliases and alts I could think of
   - fixed missing moves with the Heat property
 - bot now ignores replies to its messages
+- Details button lasts 15s, move-list buttons last 30s
 
 ## [v6.1.0] - 2026-01-09
 - Mizuumi